fixed typeorm.config, added exception catch

+++ src/account/account.service.ts
@@ -1,11 +1,18 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreateAccountDto } from './dto/create-account.dto';
 import { UpdateAccountDto } from './dto/update-account.dto';
 import { Account } from './entities/account.entity';
 import { encodePassword } from 'src/utils/bcrypt';
-import { ALREADY_REGISTERED_ERROR } from '../constants/exception.constants';
+import {
+  ACCOUNT_NOT_FOUND_ERROR,
+  ALREADY_REGISTERED_ERROR,
+} from '../constants/exception.constants';
 
 @Injectable()
 export class AccountService {
@@ -29,21 +36,35 @@ export class AccountService {
     return this.accountRepository.save(newAccount);
   }
 
-  findAll() {
-    return this.accountRepository.find();
+  async findAll() {
+    const account = await this.accountRepository.find();
+    if (!account.length) {
+      throw new NotFoundException(ACCOUNT_NOT_FOUND_ERROR);
+    }
+    return account;
   }
 
-  findOne(id: number) {
-    return this.accountRepository.findOneBy({ id });
+  async findOne(id: number) {
+    const account = await this.accountRepository.findOneBy({ id });
+    if (!account) {
+      throw new NotFoundException(ACCOUNT_NOT_FOUND_ERROR);
+    }
+    return account;
   }
 
   async update(id: number, updateAccountDto: UpdateAccountDto) {
     const account = await this.findOne(id);
+    if (!account) {
+      throw new NotFoundException(ACCOUNT_NOT_FOUND_ERROR);
+    }
     return this.accountRepository.save({ ...account, ...updateAccountDto });
   }
 
   async remove(id: number) {
     const account = await this.findOne(id);
+    if (!account) {
+      throw new NotFoundException(ACCOUNT_NOT_FOUND_ERROR);
+    }
     return this.accountRepository.remove(account);
   }
 }
